Plot Chromosome Ideogram

Draw chromosome ideogram, chromosome highlights and chromosome names. Graphic device must be initialized first. The original function was split to three new ones in order to plot chromosome highlight or names in different area. RCircos core components and graphic device must be initialized first.

Usage

RCircos.Chromosome.Ideogram.Plot(tick.interval=0)
RCircos.Draw.Chromosome.Ideogram(ideo.pos=NULL, ideo.width=NULL)
RCircos.Highligh.Chromosome.Ideogram(highlight.pos=NULL, highlight.width=NULL)
RCircos.Ideogram.Tick.Plot(tick.interval=50, track.for.ticks=3)
RCircos.Label.Chromosome.Names(chr.name.pos=NULL)

Arguments

tick.interval

Non-negative integer, the distance between two chromosome ideogram ticks in million base pairs. Set to 0 (zero) will ignore the ideogram ticks.

ideo.pos

Non-negative numeric, the distance of chromosome ideogram position from center of plot area.

ideo.width

Non-negative numeric, the width of chromosome ideogram.

highlight.pos

Non-negative numeric, the distance of chromosome highlight line position from center of plot area.

highlight.width

Non-negative integer, the width of chromosome highlight lines.

track.for.ticks

Non-negative integer, total number of tracks for ideogram tick height.

chr.name.pos

Non-negative numeric, the distance of chromosome names from center of plot area.

Details

RCircos.Chromosome.Ideogram.Plot() is the default method to plot chromosome ideogram, chromosome highlights, and chromosome names.

RCircos.Draw.Chromosome.Ideogram() plots chromosome ideogram only.

RCircos.Highligh.Chromosome.Ideogram() highlights chromosomes with different color lines only.

RCircos.Ideogram.Tick.Plot() draws chromosome ideogram ticks starting from chromosome highlight positions.

RCircos.Label.Chromosome.Names(chr.name.pos=NULL) add chromosome names for chromosome ideogram at default or customized location.

Author

Henry Zhang

Examples

if (FALSE) { # \dontrun{
library(RCircos)
data(UCSC.HG38.Human.CytoBandIdeogram)
cyto.info <- UCSC.HG38.Human.CytoBandIdeogram
RCircos.Set.Core.Components(cyto.info, chr.exclude=NULL,  
    tracks.inside=10, tracks.outside=0 )
RCircos.Set.Plot.Area()
RCircos.Chromosome.Ideogram.Plot()
} # }