Convert A Genomic Position to An Index of Circos Plot Position

Calculate the index of Circos plot positions (a set of x- and y-coordinates for a circular line) for a genomic position (chromosome name and start position). This function is for internal use only.

Usage

RCircos.Data.Point(chromosome=NULL, start.posotion=NULL)

Arguments

chromosome

Character vector, a chromosome name with prefix of "chr", e.g., "chr1".

start.posotion

Non-negative integer, start position of a genomic interval (e.g, a gene or a band) on the chromosome.

Value

An integer representing the index of RCircos plot positions.

Author

Henry Zhang