Sample Data for Gene Labels

RCircos.Gene.Label.Data contains genomic position information for 192 genes. Each row of the data contains genomic position and gene name for one gene in the order of chromosome name, start position, end position, and gene name.

Usage

data(RCircos.Gene.Label.Data)

Format

A data frame with 192 observations on the following 4 variables.

Chromosome

a factor with levels chr1 chr10 chr11 chr12 chr13 chr14 chr15 chr16 chr17 chr18 chr19 chr2 chr20 chr21 chr22 chr3 chr4 chr5 chr6 chr7 chr8 chr9 chrX

chromStart

a numeric vector

chromEnd

a numeric vector

Gene

a factor with levels ABL1 ACN9 ACVR1B AKT1 ALK APC ARHGEF2 ARID1A ASXL1 ATM ATRX BAG3 BAI3 BAP1 BCAN BCL2 BRAF BRCA1 BRCA2 CA12 CA9 CALU CARD11 CBL CCND1 CD34 CDC73 CDH1 CDK4 CDKN2A CDX2 CEACAM7 CEBPA CES3 CRLF2 CSF1R CSF3 CTNNA1 CTNNB1 CUL2 CYLD CYP1A1 DAXX DCC DES DIRAS1 DIRC2 DKK3 DLD DMBT1 DNMT3A DPYD EGF EGFR EGR3 EIF4G2 EML4 ENO1 ENO2 ERBB2 ERBB3 ERBB4 ERG EZH2 F10 FAM123B FBXW7 FGFR1 FGFR2 FGFR3 FH FHIT FLCN FLT3 FOXL2 FZR1 GATA1 GATA2 GATA3 GFAP GNA11 GNAQ GNAS GRP GSTM1 HIVEP3 HNF1A HRAS IDH1 IDH2 IFNA1 IL2 ITGB5 JAK1 JAK2 JAK3 KDR KIT KRAS KRT20 MAGI1 MAP2K4 MEN1 MET MGMT MIB1 MKI67 MLH1 MPL MSH2 MSH6 MUC1 MUC17 MUC2 MUTYH MVP MYC MYD88 NAT1 NAT2 NES NF1 NF2 NOTCH1 NOTCH2 NPM1 NRAS PBRM1 PCNA PDGFRA PDZD4 PGR PHOX2B PIK3CA PMS2 PPP2R1A PRCC PRKAR1A PTCH1 PTEN PTGS2 PTPN11 RB1 REEP5 RET RNF139 RNF2 RPS15 RUNX1 SDHA SDHAF1 SDHAF2 SDHB SDHC SDHD SELT SETD2 SLC38A1 SLC6A2 SLTM SMAD4 SMARCA4 SMARCB1 SMO SMOX SMUG1 SOCS1 SRC SST STC1 STK11 SUFU SYP TCF7L2 TET2 TFE3 TFEB TMEM97 TNFAIP3 TNFSF13 TP53 TPD52L2 TPM4 TSC1 TSC2 TSHR TYK2 VHL VIM WT1 XRCC1 ZNF135

Source

Unpublished data.