Set Colors for One Track of Heatmap

Assign colors to each value of a numeric vector for heatmap plot. The colour scale needs to be calculated from whole datasets when there are more than one data column so the minimum and maximum values for heatmap data must be provided.

Usage

RCircos.Get.Heatmap.Data.Colors(heatmap.value=NULL, min.value=NULL, max.value=NULL)

Arguments

heatmap.value

A numeric vector.

min.value

Numeric, the minimu value for heatmap scale.

max.value

Numeric,the maximum value for heatmap scale.

Value

Character vector of R color names with length same as the length of data values.

Author

Henry Zhang

Examples

if (FALSE) { # \dontrun{
library(RCircos)
data(RCircos.Heatmap.Data)
hColors <- RCircos.Get.Heatmap.Data.Colors(RCircos.Heatmap.Data[,5], min.value=-3, max.value=3)} # }