Set up Colors for Link Lines or Ribbons

Assign colors for link lines or Ribbons. The defult colors are rainbow with lenght of total link lines or ribbons. If by.chromosome is set to TRUE, red color will be used for links between positions of same chromosomes and blue color for links between different chromosomes. Users can also define their own colors by appending a column with color names to link data and give the column name as "PlotColor".

This function is for internal use only.

Usage

RCircos.Get.Link.Colors(link.data, genomic.columns=3, by.chromosome=TRUE)

Arguments

link.data

A data frame with paired genomic positions in each row. The first three columns are for chromosome names, start and end position of one genomic position followed by three columns for chromosome names, start and end position of the second genomic position.

genomic.columns

Non-negative integer, total columns that define the positions of a genomic interval (chromosome name, start and optional end positions).

by.chromosome

Logic. If true, red color will be used for links between positions of same chromosomes and blue color for links between different chromosomes. Otherwise, user defined or random colors (rainbow) will be automatically assigned for each line.

Value

A vector of color names with length same as the rows of input data.

Author

Henry Zhang