Calculate Boundary of A Data Track

Calculate plot track location based on track number, plot side, or user defined inside and outside position relative to plot area center. If all arguments are defined, user defined inside and outside position will be used. This function is mainly for internal use.

Usage

RCircos.Get.Plot.Boundary(track.num=NULL, side=NULL, 
  inside.pos=NULL, outside.pos=NULL, erase.area=FALSE)

Arguments

track.num

Non-negative integer, representing the ordinal number of the plot track where the lines will be plotted.

side

Character vector, either "in" or "out", representing the position related to chromosome ideogram. It could be NULL if both inside.pos and outside.pos are defined.

inside.pos

Non-negative numeric, inside position (relative to the centre of plot area) of the track.

outside.pos

Non-negative numeric, outside position (relative to the centre of plot area) of the track.

erase.area

Logic, if the track is going to be erased.

Value

Numeric vector of length 2, x and y coordinates of a data track.

Author

Henry Zhang