Calculate x and y Coordinates for a Bezire Curve

Calculate a quandratic Bezire curve between two points with (0,0) as control points. This function is for internal use.

Usage

RCircos.Link.Line(line.start=NULL, line.end=NULL)

Arguments

line.start

The x and y coordinates of a point where the Bezier curve starts.

line.end

The x and y coordinates of a point where the Bezier curve ends.

Author

Hongen Zhang