Cytoband Information for Rat Chromosome Ideogram

A data frame containing chromosome name, start and end position, band name, and intensity of Giemsa stains for each cytoband.

Usage

data(UCSC.Baylor.3.4.Rat.cytoBandIdeogram)

Format

A data frame with 246 observations on the following 5 variables.

Chromosome

a factor with levels chr1 chr10 chr11 chr12 chr13 chr14 chr15 chr16 chr17 chr18 chr19 chr2 chr20 chr3 chr4 chr5 chr6 chr7 chr8 chr9 chrX

ChromStart

a numeric vector

ChromEnd

a numeric vector

Band

a factor with levels p11 p12 p13 p14 p15 p16 p21 p22 q11 q12 q12.1 q12.2 q12.3 q12.4 q12.5 q13 q14 q15 q16 q21 q22 q23 q24 q25 q26 q27 q31 q32 q32.1 q32.2 q32.3 q33 q34 q35 q36 q37 q38 q41 q42 q43 q44 q45 q51 q52 q53 q54 q55

Stain

a factor with levels gneg gpos gvar

Source

http://genome.ucsc.edu