Cytoband Inforamtion for Mouse Chromosome Ideogram

A data frame containing chromosome name, start and end position, band name, and intensity of Giemsa stains for each cytoband.

Usage

data(UCSC.Mouse.GRCm38.CytoBandIdeogram)

Format

A data frame with 403 observations on the following 5 variables.

Chromosome

a factor with levels chr1 chr10 chr11 chr12 chr13 chr14 chr15 chr16 chr17 chr18 chr19 chr2 chr3 chr4 chr5 chr6 chr7 chr8 chr9 chrX chrY

ChromStart

a numeric vector

ChromEnd

a numeric vector

Band

a factor with levels qA qA1 qA1.1 qA1.2 qA1.3 qA2 qA3 qA3.1 qA3.2 qA3.3 qA4 qA5 qA5.1 qA5.2 qA5.3 qA6 qA7.1 qA7.2 qA7.3 qB qB1 qB1.1 qB1.2 qB1.3 qB2 qB2.1 qB2.2 qB2.3 qB3 qB3.1 qB3.2 qB3.3 qB4 qB5 qB5.1 qB5.2 qB5.3 qC qC1 qC1.1 qC1.2 qC1.3 qC2 qC3 qC3.1 qC3.2 qC3.3 qC4 qC5 qC6 qC7 qD qD1 qD2 qD2.1 qD2.2 qD2.3 qD3 qE qE1 qE1.1 qE1.2 qE1.3 qE2 qE2.1 qE2.2 qE2.3 qE3 qE3.1 qE3.2 qE3.3 qE4 qE5 qF qF1 qF2 qF2.1 qF2.2 qF2.3 qF3 qF4 qF5 qG1 qG1.1 qG1.2 qG1.3 qG2 qG3 qH1 qH2 qH2.1 qH2.2 qH2.3 qH3 qH4 qH5 qH6

Stain

a factor with levels gneg gpos100 gpos33 gpos66 gpos75

Source

http://genome.ucsc.edu